# k_admin/admin-web/apps/web-antd/src/api/kadmin/generated/ebook.ts
// Code generated by kadmin codegen. DO NOT EDIT.
import { request } from '../client';

export interface Ebook {
  ebookId: number;
  userId: number;
  ebookName: string;
  author: string;
  publisher: string;
  coverUrl: string;
  ebookUrl: string;
  ebookSize: number;
  ebookFormat: string;
  ebookSubject: string;
  ebookType: number;
  reviewer: string;
  ebookStatus: number;
  delFlag: number;
  createBy: string;
  createTime: string;
  updateBy: string;
  updateTime: string;
  remark: string;
}

export interface EbookPayload {
  userId: number;
  ebookName: string;
  author: string;
  publisher: string;
  coverUrl: string;
  ebookUrl: string;
  ebookSize: number;
  ebookFormat: string;
  ebookSubject: string;
  ebookType: number;
  reviewer: string;
  ebookStatus: number;
  delFlag: number;
  createBy: string;
  createTime: string;
  updateBy: string;
  updateTime: string;
  remark: string;
}

export interface EbookFilters {
  page?: number;
  pageSize?: number;
  keyword?: string;
  ebookName?: string;
  author?: string;
  publisher?: string;
  coverUrl?: string;
  ebookUrl?: string;
  ebookFormat?: string;
  ebookSubject?: string;
  reviewer?: string;
  createBy?: string;
  createTime?: string;
  updateBy?: string;
  updateTime?: string;
  remark?: string;
}

export interface PageResult<T> {
  items: T[];
  page: number;
  pageSize: number;
  total: number;
}

function queryString<T extends object>(filters: T) {
  const params = new URLSearchParams();
  Object.entries(filters).forEach(([key, value]) => {
    if (value !== undefined && value !== '') {
      params.set(key, String(value));
    }
  });
  const query = params.toString();
  return query ? `?${query}` : '';
}

export function getEbookList(filters: EbookFilters = {}) {
  return request<PageResult<Ebook>>(`/api/ebooks${queryString(filters)}`);
}

export function getEbook(id: number) {
  return request<Ebook>(`/api/ebooks/${id}`);
}

export function createEbook(payload: EbookPayload) {
  return request<Ebook>(`/api/ebooks`, {
    method: 'POST',
    body: JSON.stringify(payload),
  });
}

export function updateEbook(id: number, payload: EbookPayload) {
  return request<Ebook>(`/api/ebooks/${id}`, {
    method: 'PUT',
    body: JSON.stringify(payload),
  });
}

export function deleteEbook(id: number) {
  return request<boolean>(`/api/ebooks/${id}`, { method: 'DELETE' });
}