# k_admin/internal/kadmin/generated/ebook/types.go
// Code generated by kadmin codegen. DO NOT EDIT.

package ebook

// Permission codes for table ptmj_ebook.
const (
	ListPermission   = "system:ebook:list"
	CreatePermission = "system:ebook:create"
	UpdatePermission = "system:ebook:update"
	DeletePermission = "system:ebook:delete"
)

// Ebook is the model of table ptmj_ebook.
type Ebook struct {
	EbookId      int64  `json:"ebookId"`
	UserId       int64  `json:"userId"`
	EbookName    string `json:"ebookName"`
	Author       string `json:"author"`
	Publisher    string `json:"publisher"`
	CoverUrl     string `json:"coverUrl"`
	EbookUrl     string `json:"ebookUrl"`
	EbookSize    int64  `json:"ebookSize"`
	EbookFormat  string `json:"ebookFormat"`
	EbookSubject string `json:"ebookSubject"`
	EbookType    int64  `json:"ebookType"`
	Reviewer     string `json:"reviewer"`
	EbookStatus  int64  `json:"ebookStatus"`
	DelFlag      int64  `json:"delFlag"`
	CreateBy     string `json:"createBy"`
	CreateTime   string `json:"createTime"`
	UpdateBy     string `json:"updateBy"`
	UpdateTime   string `json:"updateTime"`
	Remark       string `json:"remark"`
}

// EbookPayload carries the writable fields of Ebook.
type EbookPayload struct {
	UserId       int64  `json:"userId"`
	EbookName    string `json:"ebookName"`
	Author       string `json:"author"`
	Publisher    string `json:"publisher"`
	CoverUrl     string `json:"coverUrl"`
	EbookUrl     string `json:"ebookUrl"`
	EbookSize    int64  `json:"ebookSize"`
	EbookFormat  string `json:"ebookFormat"`
	EbookSubject string `json:"ebookSubject"`
	EbookType    int64  `json:"ebookType"`
	Reviewer     string `json:"reviewer"`
	EbookStatus  int64  `json:"ebookStatus"`
	DelFlag      int64  `json:"delFlag"`
	CreateBy     string `json:"createBy"`
	CreateTime   string `json:"createTime"`
	UpdateBy     string `json:"updateBy"`
	UpdateTime   string `json:"updateTime"`
	Remark       string `json:"remark"`
}

// EbookFilter carries the list query parameters.
type EbookFilter struct {
	Page         int    `json:"page"`
	PageSize     int    `json:"pageSize"`
	Keyword      string `json:"keyword"`
	EbookName    string `json:"ebookName"`
	Author       string `json:"author"`
	Publisher    string `json:"publisher"`
	CoverUrl     string `json:"coverUrl"`
	EbookUrl     string `json:"ebookUrl"`
	EbookFormat  string `json:"ebookFormat"`
	EbookSubject string `json:"ebookSubject"`
	Reviewer     string `json:"reviewer"`
	CreateBy     string `json:"createBy"`
	CreateTime   string `json:"createTime"`
	UpdateBy     string `json:"updateBy"`
	UpdateTime   string `json:"updateTime"`
	Remark       string `json:"remark"`
}

// EbookPage is a paginated list result.
type EbookPage struct {
	Items    []Ebook `json:"items"`
	Total    int64   `json:"total"`
	Page     int     `json:"page"`
	PageSize int     `json:"pageSize"`
}